Part 1 of {bz: 1661}: DELETE on a deleted attachment purges the file, and now we purge when calling deleteNewAttachments
git-svn-id: https://repo.socialtext.net:8999/svn/socialtext/trunk@11085 593ce9bc-2028-0410-bb98-c566b643a948

+++ nlw/lib/Socialtext/Rest/Attachment.pm
@@ -74,7 +74,12 @@ sub DELETE {
             my $attachment = eval { $self->_get_attachment(); };
             return $self->_invalid_attachment( $rest, $@ ) if $@;
 
-            $attachment->delete( user => $rest->user );
+            if ($attachment->deleted) {
+                $attachment->purge($attachment->page);
+            }
+            else {
+                $attachment->delete( user => $rest->user );
+            }
             $rest->header( -status => HTTP_204_No_Content );
             return '';
         }
